add deno_core::ops::op_json_async / op_json_sync
